+++ api/app/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 
-from .api.api_v1.api import api_router
+from .api.api_v1.routes import api_router
 from .core.config import settings
 from .db.database import database, engine, metadata
